+++ apps/react-vite/__mocks__/zustand.ts
@@ -1,4 +1,5 @@
 import { act } from '@testing-library/react';
+import { afterEach, vi } from 'vitest';
 import * as zustand from 'zustand';
 
 const { create: actualCreate, createStore: actualCreateStore } =
@@ -18,8 +19,6 @@ const createUncurried = <T>(stateCreator: zustand.StateCreator<T>) => {
 
 // when creating a store, we get its initial state, create a reset function and add it in the set
 export const create = (<T>(stateCreator: zustand.StateCreator<T>) => {
-  console.log('zustand create mock');
-
   // to support curried version of create
   return typeof stateCreator === 'function'
     ? createUncurried(stateCreator)
@@ -37,8 +36,6 @@ const createStoreUncurried = <T>(stateCreator: zustand.StateCreator<T>) => {
 
 // when creating a store, we get its initial state, create a reset function and add it in the set
 export const createStore = (<T>(stateCreator: zustand.StateCreator<T>) => {
-  console.log('zustand createStore mock');
-
   // to support curried version of createStore
   return typeof stateCreator === 'function'
     ? createStoreUncurried(stateCreator)
