+++ apps/react-vite/src/app/index.tsx
@@ -1,24 +1,10 @@
-import { useQueryClient } from '@tanstack/react-query';
-import { useMemo } from 'react';
-import { RouterProvider } from 'react-router-dom';
+import { AppProvider } from './provider';
+import { AppRouter } from './router';
 
-import { AppProvider } from './main-provider';
-import { createRouter } from './routes';
-
-const AppRouter = () => {
-  const queryClient = useQueryClient();
-
-  const router = useMemo(() => createRouter(queryClient), [queryClient]);
-
-  return <RouterProvider router={router} />;
-};
-
-function App() {
+export const App = () => {
   return (
     <AppProvider>
       <AppRouter />
     </AppProvider>
   );
-}
-
-export default App;
+};
