+++ apps/react-vite/src/lib/react-query.ts
@@ -1,8 +1,4 @@
-import {
-  UseMutationOptions,
-  DefaultOptions,
-  QueryClient,
-} from '@tanstack/react-query';
+import { UseMutationOptions, DefaultOptions } from '@tanstack/react-query';
 
 export const queryConfig = {
   queries: {
@@ -12,10 +8,6 @@ export const queryConfig = {
     staleTime: 1000 * 60,
   },
 } satisfies DefaultOptions;
-
-export const queryClient = new QueryClient({
-  defaultOptions: queryConfig,
-});
 
 export type ApiFnReturnType<FnType extends (...args: any) => Promise<any>> =
   Awaited<ReturnType<FnType>>;
