+++ apps/react-vite/src/main.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 
 import './index.css';
-import App from './app';
+import { App } from './app';
 import { enableMocking } from './testing/mocks';
 
 const root = document.getElementById('root');
