+++ apps/react-vite/src/testing/setup-tests.ts
@@ -1,6 +1,5 @@
 import '@testing-library/jest-dom/vitest';
 
-import { queryClient } from '@/lib/react-query';
 import { initializeDb, resetDb } from '@/testing/mocks/db';
 import { server } from '@/testing/mocks/server';
 
@@ -25,5 +24,4 @@ beforeEach(() => {
 afterEach(() => {
   server.resetHandlers();
   resetDb();
-  queryClient.clear();
 });
