+++ docs/project-structure.md
@@ -9,7 +9,8 @@ src
 |   |
 |   +-- routes        # application routes / can also be called pages
     +-- app.tsx       # main application component
-    +-- app-provider  # application provider that wraps the entire application with global providers
+    +-- provider.tsx  # application provider that wraps the entire application with different global providers
+    +-- router.tsx    # application router configuration
 +-- assets            # assets folder can contain all the static files such as images, fonts, etc.
 |
 +-- components        # shared components used across the entire application
@@ -55,6 +56,10 @@ src/features/awesome-feature
 
 NOTE: You don't need all of these folders for every feature. Only include the ones that are necessary for the feature.
 
+In some cases it might be more practical to keep all api calls outside of the feature folders in a dedicated `api` folder where all API calls are defined. This can be useful if you have a lot of shared api calls between features.
+
+````sh
+
 In the past, it was recommended to use barrel files to export all the files from a feature. However, it can cause issues for Vite to do tree shaking and can lead to performance issues. Therefore, it is recommended to import the files directly.
 
 It might not be a good idea to import across the features. Instead, compose different features at the application level. This way, you can ensure that each feature is independent which makes the codebase less convoluted.
@@ -98,7 +103,7 @@ To forbid cross-feature imports, you can use ESLint:
         ],
     },
 ],
-```
+````
 
 You might also want to enforce unidirectional codebase architecture. This means that the code should flow in one direction, from shared parts of the code to the application (shared -> features -> app). This is a good practice to follow as it makes the codebase more predictable and easier to understand.
